refactor: extract crypto + safety helpers out of settings.py
Per maintainer direction ("settings.py should ideally just read in from
env, maybe with a couple of (very) lightweight functions"), split the
non-config-read helpers out of settings.py and into modules that own
the responsibility:

New: ``memory.common.crypto`` (a leaf module — depends on nothing in
this codebase, so both ``settings`` and any future caller can import
it without circularity):
  - ``derive_transfer_token_secret(master_key, salt)``
  - ``TRANSFER_TOKEN_SECRET_HKDF_INFO``
  The function now takes ``salt`` as an explicit argument rather than
  reading ``SECRETS_ENCRYPTION_SALT`` from settings module scope —
  settings.py passes the salt when computing
  ``TRANSFER_TOKEN_SECRET`` at module load.

New: ``memory.api.safety`` (lives next to its only caller, ``api/app.py``):
  - ``is_loopback_url(url)``
  - ``validate_disable_auth_safety()``
  The validator reads from ``settings.X`` directly; settings does not
  depend on safety in either direction. ``app.py``'s startup line goes
  from ``settings.validate_disable_auth_safety()`` to
  ``validate_disable_auth_safety()`` imported from ``api.safety``.

``settings.py`` is back to a config leaf:
  - ``boolean_env`` / ``secret_env`` env-reader helpers
  - module-level ``os.getenv`` reads + path config
  - the 3-way ``TRANSFER_TOKEN_SECRET`` assignment (explicit env →
    HKDF-derive via ``crypto`` → None), which still computes once at
    import time per the operator-facing contract.

The ``cryptography.hazmat`` and ``urllib.parse.urlparse`` imports are
no longer pulled into ``settings.py`` — modules that import settings
for config now incur no transitive crypto/url-parser dependency.

Tests split to match:
  - ``tests/memory/common/test_settings.py`` keeps only the
    ``secret_env`` cases (the only behaviour settings.py still owns).
  - ``tests/memory/common/test_crypto.py`` (new) covers
    ``derive_transfer_token_secret`` — determinism, distinct-from-master,
    hex shape, versioned info, salt-dependence. Salt is passed
    explicitly in each test rather than reaching into ``settings.SECRETS_ENCRYPTION_SALT``.
  - ``tests/memory/api/test_safety.py`` (new) covers
    ``validate_disable_auth_safety`` (all of the original DISABLE_AUTH
    test cases) plus ``is_loopback_url`` (broken out as a unit test
    since it's now a public helper in its own module).
